RP2040_Slave.py: Major rewrite.
First of all, the API is changed to a higher-level one, where only
3 methods are needed for normal operation:

  poll()
  do_receive()
  do_respond()

The poll() is optimized for fast execution, whereas the two others
are _generalised_ to a much more high-level API.
This API allows a much cleaner and smaller client code, where the
fast poll() gives more room for useful computation when the bus is
idle.

In addition there are some cleanups of unsued constants, and more
PEP-8 compliant naming conventions of identifiers.